Configure apt-get to see GCP's pre-built ruby binary packages and preinstall only one ruby using those packages.
This should reduce the base image size considerably, but at the cost of eliminating some preinstalled rubies. Therefore, this should not be merged until the gcloud sdk generates Dockerfiles that use the pre-built binary packages to speed up ruby installations. Note that we are also changing the default ruby version from 2.3.0 to 2.3.1.