Provide a Dockerfile for obsolete google/ruby images.
This Dockerfile creates a disabled image, displaying an error on image and container creation that points users to the correct base images to use.
Once it is in place, we will switch the auto-build for the obsolete google/ruby, google/ruby-runtime, and google/ruby-hello images to point to this Dockerfile, disabling those images. We may, at some point later, actually delete the images altogether.